codacy

+++ backend/tests/db/db-handler.js
@@ -16,7 +16,7 @@ module.exports.connect = async () => {
     };
 
     await mongoose.connect(uri, mongooseOpts);
-}
+};
 
 /**
  * Drop database, close the connection and stop mongod.
@@ -25,7 +25,7 @@ module.exports.closeDatabase = async () => {
     await mongoose.connection.dropDatabase();
     await mongoose.connection.close();
     await mongod.stop();
-}
+};
 
 /**
  * Remove all the data for all db collections.
@@ -34,7 +34,10 @@ module.exports.clearDatabase = async () => {
     const collections = mongoose.connection.collections;
 
     for (const key in collections) {
-        const collection = collections[parseInt(key, 10)];
-        await collection.deleteMany();
+        if ({}.hasOwnProperty.call(foo, key)) {
+            const collection = key;
+            await collection.deleteMany();
+        }
+
     }
-}
+};
